% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apastylr.R
\name{apastylr}
\alias{apastylr}
\title{Create a data frame with APA rules formated coefficient informations 
of a linear model}
\usage{
apastylr(model, statistics = "T")
}
\arguments{
\item{model}{A linear model created using lm().}

\item{statistics}{Reports student's t ("T") or Fisher's F ("F").}
}
\description{
Create a data frame with APA rules formated coefficient informations 
of a linear model
}

